// wardn/frontend/lib/api/generated/model/agentRunStepRead.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunStepReadPayload } from './agentRunStepReadPayload';

export interface AgentRunStepRead {
  agentRunId: string;
  createdAt: string;
  id: string;
  mcpToolInvocationId?: string | null;
  payload?: AgentRunStepReadPayload;
  sequence: number;
  status: string;
  stepType: string;
  title: string;
  updatedAt: string;
}